Address comments: change the way at signal redirection.
