+++ src/v2/view-builder/views/idp/AuthenticatorIdPVerifyView.js
@@ -1,8 +1,8 @@
-import { createCallout, loc } from 'okta';
-import { BaseForm } from '../../internals';
+import { loc } from 'okta';
+import BaseIdPAuthenticatorBody from './BaseIdPAuthenticatorBody';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 
-const Body = BaseForm.extend({
+const Body = BaseIdPAuthenticatorBody.extend({
 
   title() {
     const displayName = this.options.appState.getAuthenticatorDisplayName();
@@ -12,29 +12,6 @@ const Body = BaseForm.extend({
   subtitle() {
     const displayName = this.options.appState.getAuthenticatorDisplayName();
     return loc('oie.idp.challenge.description', 'login', [displayName]);
-  },
-
-  showMessages() {
-    // IdP Authenticator error messages are not form errors
-    // Parse and display them here.
-    const messages = this.options.appState.get('messages') || {};
-    if (Array.isArray(messages.value)) {
-      this.add('<div class="ion-messages-container”></div>', '.o-form-error-container');
-
-      messages
-        .value
-        .forEach(messagesObj => {
-          const msg = messagesObj.message;
-          if (messagesObj.class === 'ERROR') {
-            this.add(createCallout({
-              content: msg,
-              type: 'error',
-            }), '.o-form-error-container');
-          } else {
-            this.add(`<p>${msg}</p>`, '.ion-messages-container');
-          }
-        });
-    }
   },
 
   save() {
