+++ test/testcafe/spec/Interact_spec.js
@@ -7,7 +7,7 @@ import xhrWellKnownResponse from '../../../playground/mocks/data/oauth2/well-kno
 import xhrInteractResponse from '../../../playground/mocks/data/oauth2/interact.json';
 import xhrIdentify from '../../../playground/mocks/data/idp/idx/identify';
 
-const expectIdentifyView =  {
+const expectIdentifyView = {
   controller: 'primary-auth',
   formName: 'identify',
 };
@@ -38,7 +38,7 @@ const mockCrypto = ClientFunction(() => {
 
   if (typeof window.crypto.subtle === 'undefined') {
     window.crypto.subtle = {
-      digest: function() { 
+      digest: function() {
         return Promise.resolve(65);
       }
     };
@@ -71,7 +71,9 @@ const errorMock = RequestMock()
   .onRequestTo('http://localhost:3000/oauth2/default/.well-known/openid-configuration')
   .respond(xhrWellKnownResponse, 200)
   .onRequestTo('http://localhost:3000/oauth2/default/v1/interact')
-  .respond(xhrServerError, 400);
+  .respond(xhrServerError, 400)
+  .onRequestTo('http://localhost:3000/idp/idx/introspect')
+  .respond(xhrIdentify);
 
 
 const requestLogger = RequestLogger(
@@ -115,7 +117,7 @@ async function setup(t, options = {}) {
   // Re-render the widget for interaction code flow
   await mockCrypto();
   await rerenderWidget({
-    stateToken: undefined, 
+    stateToken: undefined,
     clientId: 'fake',
     redirectUri: 'http://doesnot-matter',
     useInteractionCodeFlow: true,
